refactor: :recycle: Change function location.
